requirements.txt
google-api-python-client >= 2.119.0
google-auth >= 2.26.2
langchain >= 0.1.0
streamlit-oauth >= 0.1.8
streamlit >= 1.31.1
langchain-google-genai >= 0.0.6
pillow >= 10.0.1
extra-streamlit-components >= 0.1.60
streamlit-extras >= 0.4.0
faiss-cpu >= 1.7.4
streamlit-js-eval >= 0.1.5